+++ src/config/Save.zig
@@ -1,2 +1,2 @@
 user: ?[]const u8 = null,
-session_index: ?u64 = null,
+session_index: ?usize = null,
